contracts/fix: Handle overflows in rewards calculations
The calculation of rewards has the potential to cause overflows at high
enough staked amounts and alphas. This has been accounted for within
reasonable limits. Tests have been added for the boundary conditions,
but those boundary conditions which are not reasonably likely to happen
have been ignored. For example, one boundary condition is that
block.timestamp exceeds 2e59, which is more than 3 times the age of the
universe. We have no intention of accomodating such unreasonable
conditions.