contracts/chore: Rename staker to user in staking contract
The word `staker` is a little contrived and is hard to use in
conjunction with other terms in the contract. For eg. a variable named
`stakerStakedAmt` (which refers to the amount staked by a user, as
compared to the total staked amount) is not as elegant.

So the word `staker` was replaced throughout by the word `user`, which
has more natural connotations