contracts/refactor: Update alpha directly within `_updateReward`
Earlier alpha was being updated in its own function `_updateAlpha`. This
was made a view function `_calculateAccruedAlpha`, and the actual
updating of alpha was done within the `_updateReward` function. This
makes the logic within it clean - It calls other view functions for the
calculation, and handles the events and storage variable updates within
itself.