web-app/feat: Add root error boundary
Apart from this, the error returned from '/login' API was modified to
conform to the format expected by the error boundary. Namely the error
data was returned as a single string instead of an object. We can log
additional details separately in our logger. This is only for client
facing purposes