web-app/refactor: Standardize FSC Handler class methods for provider/signer
Some methods of the smart contract require signers. Others require providers.
We need a standardized structure to approach how to handle and store the
contract instances, and how to read from them.

What we have decided is that each public function has the ability to
decide at it's level, whether it needs providers or signers. Subsequently,
it calls private functions which can be marked as a `query` or `mutation`,
which signals whether we need to pass in a contract handler with a provider
or signer.

This approach is also additive to our larger application data handling
structure. Data is cached at the RTK Query level (application level). So
there is no need to cache low level calls at our contract handler level.