web-app/feat: Add cooldown checking to _Finthetix_ contract interactions
We check if the contract is cooling down, before trying to interact with
it.  This saves us gas by preventing reverts, and also improves the UX.

This is currently done only in a guard check format. This needs to be
displayed prominently on the user's dashboard as well, and also needs to
be handled better by the frontend