web-app/feat: Fetch cooldown time from backend contract
The cooldown time is fetched as part of the `Finthetix Status`, from the
contract via RTK Query endpoint. It is refreshed whenever user tries some
interaction. Good thing to note is that if the cooldown check fails when
user tries to interact, that interaction will also trigger a refetch of
`FinthetixStatus`, causing the _Cooling Down_ banner to be shown correctly.