contracts/test: Fix address conflicts in `FinthetixStakingContract` tests
In our tests we now make sure that the randomly generated addresses don't
conflict with the address of the staking contract and associated tokens.
We also make sure that if there are multiple addresses being used inside
the test, that those don't overlap as well.

This bug in the tests were making the fuzz tests fail occasionally, since
there's only a small chance that the randomly generated addresses overlap