contracts/test: Prevent conflicts in static burner addresses
For some of the tests we don't fuzz for random addresses, but rather use
static addresses to impersonate a user. However, we never considered the
fact that these addresses could conflict with the `Finthetix` addresses
as well. So in order to make sure this is not the case, we explicitly
check this in the `setup()` function.

The addresses of the `Finthetix` contracts depend on some random seed.
So we can trust that for a given condition of the codebase, the addresses
will be consistent. Hence we are okay with explicitly failing these tests
in case there is a conflict, since in such a condition we can manually
alter these burner addresses (effectively changing the seed)