chore: Allow infra folder to read from other folders
By doing so we can import the env schemas from other projects.

In order to support the same, we needed to change the build process for
the infra repo. This involved getting substituting `ts-node` with `tsx`.

As a side task, the build and watch scripts were removed from package.json
as they are [redundant](https://github.com/aws/aws-cdk/discussions/19516#discussioncomment-2456520)