web-app/fix: Close modal when user rejects transaction
When the user rejects a staking/unstaking transaction from metamask while
the modal is open, it causes the user data to refresh. While the data is
refreshing, the modal is empty. This is unwanted behaviour.  To fix this,
we close the modal automatically when user data is loading.

The loading spinner is anyway shown on the card (on Staked Amount number)
when user data is loading, and the Stake/Unstake buttons are disabled.