web-app/refactor: Move dappInfo and chainInfo schemas into dedicated files
A cyclical dependency was being created, as the zod schema for
dappInfo/chainInfo is used in the `process.env` parser, but the `process.env`
parser is used in the dappInfo/chainInfo getter. The problem was easily
solved by moving the schema and the getter into their own dedicated files.

All imports to these were also updated