web-app/refactor: Use chainInfo and dappInfo from env, not root loader
Earlier, the chainInfo and dappInfo were passed in as parameters to the
RTK query layer. This is problematic in the following ways:
  - These parameters, while required information, are not pertinent to
    the request/endpoint in question. Therefore, it seems odd that we are
    passing this in as information.

  - We had to read the root loader data at multiple levels of the component
    tree.  While this doesn't incur multiple requests (loaders only run
    once, regardless of how many times you read their data), the data in
    question (dappInfo, chainInfo) actually resembles environment variables,
    and it makes sense to have them available globally, and not have to
    fetch them explicitly.

  - This is a lot of repeated code

We fixed this, by removing said params, and instead reading the chainInfo
& dappInfo from the parsed browser env. We do this directly in the
`FinthetixStakingContractHandler`, which removes a lot of repeated code.

The metamask handler generation was also moved into the
`FinthetixStakingContractHandler.make` function, thereby removing even
more repeated code