web-app/feat: Try to use User's RPC first
Trying to use the network already added by the user, reduces a lot of
friction and reduces the trust required to use the dapp. We first try to
switch to user's RPC on the correct network. If they reject, that's that.
Else if they have no RPC, we suggest our own. That way we also save on
our RPC requests.