web-app/fix: Use svh instead of dvh
Using dvh causes a small amount of jittering when the address bar shrinks
when the user scrolls. This can be perceived as a bug. To avoid this
perception, we use small viewport height (svh) instead, so that user
gets the full experience (full screen) even when the address bar is shown